reorganize pages folder structure
