+++ src/pages/Auth/ResetPasswordPage.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { useParams, useNavigate } from "react-router-dom";
-import api from "../services/api";
+import api from "../../services/api";
 import { toast } from "react-toastify";
 
 const ResetPasswordPage = () => {
