+++ src/pages/Home/HomePage.tsx
@@ -1,0 +1,17 @@
+import Hero from "../../components/Home/Hero";
+import Features from "../../components/Home/Features";
+import CarouselInfo from "../../components/Home/CarouselInfo";
+import Footer from "../../components/Layout/Footer";
+
+const Home = () => {
+  return (
+    <>
+      <Hero />
+      <Features />
+      <CarouselInfo />
+      <Footer />
+    </>
+  );
+};
+
+export default Home;
